Moved Writer's interfaces under "storage.io" package
