+++ hudi-common/src/main/java/org/apache/hudi/common/model/HoodieRecord.java
@@ -21,10 +21,10 @@ package org.apache.hudi.common.model;
 import org.apache.avro.Schema;
 import org.apache.avro.generic.GenericRecord;
 import org.apache.hudi.common.config.TypedProperties;
-import org.apache.hudi.common.io.storage.HoodieFileWriter;
 import org.apache.hudi.common.util.CollectionUtils;
 import org.apache.hudi.common.util.Option;
 import org.apache.hudi.common.util.collection.Pair;
+import org.apache.hudi.io.storage.HoodieFileWriter;
 
 import java.io.IOException;
 import java.io.Serializable;
